Moved RockMod class to server dir.
